% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{csu_energy_palette}
\alias{csu_energy_palette}
\title{CSU Energy Palette}
\usage{
csu_energy_palette()
}
\value{
Named character vector of hex codes
}
\description{
Returns the official CSU energy / secondary brand colors.
Each color is named according to CSU branding guidelines.
}
