% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{csu_pal}
\alias{csu_pal}
\title{Get Colors from a CSU Palette}
\usage{
csu_pal(palette = c("primary", "energy"), n = NULL)
}
\arguments{
\item{palette}{Character: "primary" or "energy"}

\item{n}{Number of colors to return. If NULL, returns all colors in the palette.}
}
\value{
Character vector of hex codes
}
\description{
Interpolate colors from a CSU palette.
}
\examples{
csu_pal("primary")
csu_pal("energy", 10)
}
