% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{csu_primary_palette}
\alias{csu_primary_palette}
\title{CSU Primary Palette}
\usage{
csu_primary_palette()
}
\value{
Named character vector of hex codes
}
\description{
Returns the official CSU primary brand colors.
Each color is named according to CSU branding guidelines.
}
