% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{show_csu_palette}
\alias{show_csu_palette}
\title{Display a CSU Palette}
\usage{
show_csu_palette(palette = c("primary", "energy"))
}
\arguments{
\item{palette}{Character: "primary" or "energy"}
}
\description{
Show a palette as a horizontal color strip with names.
}
