% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{csu_pal}
\alias{csu_pal}
\title{CSU Palettes}
\usage{
csu_pal(palette = c("primary", "energy"), n = NULL)
}
\arguments{
\item{palette}{Which palette to use: "primary" or "energy".}

\item{n}{Number of colors to return. If \code{NULL}, returns all colors in the palette.}
}
\value{
A vector of hex color codes.
}
\description{
Access and interpolate CSU color palettes.
}
\examples{
csu_pal("primary")
csu_pal("energy", 10)
}
